+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0).
 
 ## Unreleased
 
-Nothing.
+### Fixed
+
+- Properly handle parsing of invalid config instead of bubbling up internal
+  exceptions.
 
 ## [2.0.2](https://github.com/trallnag/filter-pre-commit-hooks/compare/v2.0.1...v2.0.2) / 2025-02-19